Construct an increment transaction
Starting with the off-chain code to create a transaction for the Open ->
Open transition using the 'Increment' redeemer. While this is correctly
spending the committed UTxOs and recreating a Head output, it is
not (yet) using a sparse merkle tree (SMT) to record the added UTxO
correctly on-chain (and proof to the script that it was committing
correctly).