Add committedRefs to increment redeemer
This is how the increment validator should work. It would point to the
spent UTxO that are recorded and to be added to the state.

This is currently not possible though as the known UTxO at the time of
observation does not include those "external" UTxO (see TODO). If we
would have the transaction inputs resolved (which must be available at
the time of observation), then we could observe this correctly.